// src/final-project/models/enums.ts
export enum Priority {
  Low = "LOW",
  Medium = "MEDIUM",
  High = "HIGH",
  Urgent = "URGENT",
}

export enum TaskStatus {
  Pending = "PENDING",
  InProgress = "IN_PROGRESS",
  Completed = "COMPLETED",
}

export enum Category {
  Work = "WORK",
  Personal = "PERSONAL",
  Shopping = "SHOPPING",
  Health = "HEALTH",
  Other = "OTHER",
}